Refactor cleaning pkg
